+++ src/main/java/com/sparta/engineering72/Settings/Settings.java
@@ -1,5 +1,7 @@
 package com.sparta.engineering72.Settings;
 
+import java.math.BigInteger;
+
 public class Settings {
-    public static final int MAX_COUNT_THRESHOLD = 10000;
+    public static final BigInteger MAX_COUNT_THRESHOLD = BigInteger.valueOf(10000);
 }
